Revert "Navbar & Hamburger added"
