+++ package.json
@@ -22,7 +22,6 @@
     "react-photo-gallery": "^8.0.0",
     "react-router-dom": "^6.6.2",
     "react-scripts": "^5.0.1",
-    "styled-components": "^5.3.6",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
