--- src/components/Navbar/css/style.css
@@ -1,43 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Sen:wght@400;700;800&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Manjari:wght@700&display=swap');
-
-
-html {
-    scroll-behavior: smooth;
-}
-
-Nav {
-    background: #232536;
-    margin: 0;
-    padding: 0;
-    color: #fff;
-
-    width: 100%;
-    height: 63px;
-    padding-left: 18px;
-
-    text-decoration: none;
-    font-family: inter;
-    font-weight: 400px;
-    font-size: 16px;
-}
-
-a:visited {
-    color: #fff;
-}
-
-a {
-    text-decoration: none;
-}
-
-.RKH {
-    font-family: Manjari;
-    font-weight: 700;
-    font-size: 18px;
-    padding-left: 40px;
-}
-
-.logo{
-    margin-top: 13px;
-}
